@@ -3,19 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.2",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.0",
-    "axios": "^0.27.2",
-    "react": "^17.0.2",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.2",
-    "react-router-dom": "^5.3.0",
-    "web-vitals": "^2.1.4"
+    "@reduxjs/toolkit": "^2.2.3",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.6",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.6.8",
+    "react": "^18.3.1",
+    "react-cookie": "^7.1.4",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.23.0",
+    "web-vitals": "^3.5.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.2.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-react": "^7.34.1",
+    "globals": "^15.1.0",
+    "prettier": "3.2.5",
     "react-scripts": "5.0.1"
   },
   "scripts": {
@@ -27,7 +33,8 @@
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest"
+      "react-app/jest",
+      "prettier"
     ]
   },
   "browserslist": {
