@@ -1,64 +1,67 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useCookies } from "react-cookie";
-import { useHistory, useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import { Header } from "../components/Header";
 import { url } from "../const";
 import "./editList.css";
 
 export const EditList = () => {
-  const history = useHistory();
+  const history = useNavigate();
   const { listId } = useParams();
   const [title, setTitle] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
-  const [ cookies ] = useCookies();
+  const [cookies] = useCookies();
   const handleTitleChange = (e) => setTitle(e.target.value);
   const onUpdateList = () => {
     const data = {
-      title: title
-    }
+      title: title,
+    };
 
-    axios.put(`${url}/lists/${listId}`, data, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then(() => {
-      history.push("/");
-    })
-    .catch((err) => {
-    setErrorMessage(`更新に失敗しました。 ${err}`);
-    })
-  }
+    axios
+      .put(`${url}/lists/${listId}`, data, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then(() => {
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`更新に失敗しました。 ${err}`);
+      });
+  };
 
   const onDeleteList = () => {
-    axios.delete(`${url}/lists/${listId}`, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then(() => {
-      history.push("/");
-    })
-    .catch((err) => {
-      setErrorMessage(`削除に失敗しました。${err}`);
-    })
-  }
+    axios
+      .delete(`${url}/lists/${listId}`, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then(() => {
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`削除に失敗しました。${err}`);
+      });
+  };
 
   useEffect(() => {
-    axios.get(`${url}/lists/${listId}`, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then((res) => {
-      const list = res.data;
-      setTitle(list.title);
-    })
-    .catch((err) => {
-      setErrorMessage(`リスト情報の取得に失敗しました。${err}`);
-    })
-  }, [])
+    axios
+      .get(`${url}/lists/${listId}`, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then((res) => {
+        const list = res.data;
+        setTitle(list.title);
+      })
+      .catch((err) => {
+        setErrorMessage(`リスト情報の取得に失敗しました。${err}`);
+      });
+  }, []);
 
   return (
     <div>
@@ -67,12 +70,31 @@ export const EditList = () => {
         <h2>リスト編集</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="edit-list-form">
-          <label>タイトル</label><br />
-          <input type="text" className="edit-list-title" value={title} onChange={handleTitleChange} /><br />
-          <button type="button" className="delete-list-button" onClick={onDeleteList}>削除</button>
-          <button type="button" className="edit-list-button" onClick={onUpdateList}>更新</button>
+          <label>タイトル</label>
+          <br />
+          <input
+            type="text"
+            className="edit-list-title"
+            value={title}
+            onChange={handleTitleChange}
+          />
+          <br />
+          <button
+            type="button"
+            className="delete-list-button"
+            onClick={onDeleteList}
+          >
+            削除
+          </button>
+          <button
+            type="button"
+            className="edit-list-button"
+            onClick={onUpdateList}
+          >
+            更新
+          </button>
         </form>
       </main>
     </div>
-  )
-}
+  );
+};
