@@ -3,11 +3,11 @@ import { Header } from "../components/Header";
 import axios from "axios";
 import { useCookies } from "react-cookie";
 import { url } from "../const";
-import { useHistory, useParams } from "react-router-dom";
-import "./editTask.css"
+import { useNavigate, useParams } from "react-router-dom";
+import "./editTask.css";
 
 export const EditTask = () => {
-  const history = useHistory();
+  const history = useNavigate();
   const { listId, taskId } = useParams();
   const [cookies] = useCookies();
   const [title, setTitle] = useState("");
@@ -18,57 +18,60 @@ export const EditTask = () => {
   const handleDetailChange = (e) => setDetail(e.target.value);
   const handleIsDoneChange = (e) => setIsDone(e.target.value === "done");
   const onUpdateTask = () => {
-    console.log(isDone)
+    console.log(isDone);
     const data = {
       title: title,
       detail: detail,
-      done: isDone
-    }
+      done: isDone,
+    };
 
-    axios.put(`${url}/lists/${listId}/tasks/${taskId}`, data, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then((res) => {
-      console.log(res.data)
-      history.push("/");
-    })
-    .catch((err) => {
-      setErrorMessage(`更新に失敗しました。${err}`);
-    })
-  }
+    axios
+      .put(`${url}/lists/${listId}/tasks/${taskId}`, data, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then((res) => {
+        console.log(res.data);
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`更新に失敗しました。${err}`);
+      });
+  };
 
   const onDeleteTask = () => {
-    axios.delete(`${url}/lists/${listId}/tasks/${taskId}`, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then(() => {
-      history.push("/");
-    })
-    .catch((err) => {
-      setErrorMessage(`削除に失敗しました。${err}`);
-    })
-  }
+    axios
+      .delete(`${url}/lists/${listId}/tasks/${taskId}`, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then(() => {
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`削除に失敗しました。${err}`);
+      });
+  };
 
   useEffect(() => {
-    axios.get(`${url}/lists/${listId}/tasks/${taskId}`, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then((res) => {
-      const task = res.data
-      setTitle(task.title)
-      setDetail(task.detail)
-      setIsDone(task.done)
-    })
-    .catch((err) => {
-      setErrorMessage(`タスク情報の取得に失敗しました。${err}`);
-    })
-  }, [])
+    axios
+      .get(`${url}/lists/${listId}/tasks/${taskId}`, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then((res) => {
+        const task = res.data;
+        setTitle(task.title);
+        setDetail(task.detail);
+        setIsDone(task.done);
+      })
+      .catch((err) => {
+        setErrorMessage(`タスク情報の取得に失敗しました。${err}`);
+      });
+  }, []);
 
   return (
     <div>
@@ -77,18 +80,60 @@ export const EditTask = () => {
         <h2>タスク編集</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="edit-task-form">
-          <label>タイトル</label><br />
-          <input type="text" onChange={handleTitleChange} className="edit-task-title" value={title} /><br />
-          <label>詳細</label><br />
-          <textarea type="text" onChange={handleDetailChange} className="edit-task-detail" value={detail} /><br />
+          <label>タイトル</label>
+          <br />
+          <input
+            type="text"
+            onChange={handleTitleChange}
+            className="edit-task-title"
+            value={title}
+          />
+          <br />
+          <label>詳細</label>
+          <br />
+          <textarea
+            type="text"
+            onChange={handleDetailChange}
+            className="edit-task-detail"
+            value={detail}
+          />
+          <br />
           <div>
-            <input type="radio" id="todo" name="status" value="todo" onChange={handleIsDoneChange} checked={isDone === false ? "checked" : ""} />未完了
-            <input type="radio" id="done" name="status" value="done" onChange={handleIsDoneChange} checked={isDone === true ? "checked" : ""} />完了
+            <input
+              type="radio"
+              id="todo"
+              name="status"
+              value="todo"
+              onChange={handleIsDoneChange}
+              checked={isDone === false ? "checked" : ""}
+            />
+            未完了
+            <input
+              type="radio"
+              id="done"
+              name="status"
+              value="done"
+              onChange={handleIsDoneChange}
+              checked={isDone === true ? "checked" : ""}
+            />
+            完了
           </div>
-          <button type="button" className="delete-task-button" onClick={onDeleteTask}>削除</button>
-          <button type="button" className="edit-task-button" onClick={onUpdateTask}>更新</button>
+          <button
+            type="button"
+            className="delete-task-button"
+            onClick={onDeleteTask}
+          >
+            削除
+          </button>
+          <button
+            type="button"
+            className="edit-task-button"
+            onClick={onUpdateTask}
+          >
+            更新
+          </button>
         </form>
       </main>
     </div>
-  )
-}
+  );
+};
