@@ -2,33 +2,34 @@ import React, { useState } from "react";
 import { useCookies } from "react-cookie";
 import axios from "axios";
 import { Header } from "../components/Header";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { url } from "../const";
 import "./newList.css";
 
 export const NewList = () => {
   const [cookies] = useCookies();
-  const history = useHistory();
+  const history = useNavigate();
   const [title, setTitle] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
   const handleTitleChange = (e) => setTitle(e.target.value);
   const onCreateList = () => {
     const data = {
-      title: title
-    }
+      title: title,
+    };
 
-    axios.post(`${url}/lists`, data, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then(() => {
-      history.push("/");
-    })
-    .catch((err) => {
-      setErrorMessage(`リストの作成に失敗しました。${err}`);
-    })
-  }
+    axios
+      .post(`${url}/lists`, data, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then(() => {
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`リストの作成に失敗しました。${err}`);
+      });
+  };
 
   return (
     <div>
@@ -37,11 +38,23 @@ export const NewList = () => {
         <h2>リスト新規作成</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="new-list-form">
-          <label>タイトル</label><br />
-          <input type="text" onChange={handleTitleChange} className="new-list-title" /><br />
-          <button type="button" onClick={onCreateList} className="new-list-button">作成</button>
+          <label>タイトル</label>
+          <br />
+          <input
+            type="text"
+            onChange={handleTitleChange}
+            className="new-list-title"
+          />
+          <br />
+          <button
+            type="button"
+            onClick={onCreateList}
+            className="new-list-button"
+          >
+            作成
+          </button>
         </form>
       </main>
     </div>
-  )
-}
+  );
+};
