@@ -3,8 +3,8 @@ import { useCookies } from "react-cookie";
 import axios from "axios";
 import { url } from "../const";
 import { Header } from "../components/Header";
-import "./newTask.css"
-import { useHistory } from "react-router-dom";
+import "./newTask.css";
+import { useNavigate } from "react-router-dom";
 
 export const NewTask = () => {
   const [selectListId, setSelectListId] = useState();
@@ -13,7 +13,7 @@ export const NewTask = () => {
   const [detail, setDetail] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
   const [cookies] = useCookies();
-  const history = useHistory();
+  const history = useNavigate();
   const handleTitleChange = (e) => setTitle(e.target.value);
   const handleDetailChange = (e) => setDetail(e.target.value);
   const handleSelectList = (id) => setSelectListId(id);
@@ -24,33 +24,35 @@ export const NewTask = () => {
       done: false,
     };
 
-    axios.post(`${url}/lists/${selectListId}/tasks`, data, {
+    axios
+      .post(`${url}/lists/${selectListId}/tasks`, data, {
         headers: {
-          authorization: `Bearer ${cookies.token}`
-        }
-    })
-    .then(() => {
-      history.push("/");
-    })
-    .catch((err) => {
-      setErrorMessage(`タスクの作成に失敗しました。${err}`);
-    })
-  }
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then(() => {
+        history.push("/");
+      })
+      .catch((err) => {
+        setErrorMessage(`タスクの作成に失敗しました。${err}`);
+      });
+  };
 
   useEffect(() => {
-    axios.get(`${url}/lists`, {
-      headers: {
-        authorization: `Bearer ${cookies.token}`
-      }
-    })
-    .then((res) => {
-      setLists(res.data)
-      setSelectListId(res.data[0]?.id)
-    })
-    .catch((err) => {
-      setErrorMessage(`リストの取得に失敗しました。${err}`);
-    })
-  }, [])
+    axios
+      .get(`${url}/lists`, {
+        headers: {
+          authorization: `Bearer ${cookies.token}`,
+        },
+      })
+      .then((res) => {
+        setLists(res.data);
+        setSelectListId(res.data[0]?.id);
+      })
+      .catch((err) => {
+        setErrorMessage(`リストの取得に失敗しました。${err}`);
+      });
+  }, []);
 
   return (
     <div>
@@ -59,19 +61,44 @@ export const NewTask = () => {
         <h2>タスク新規作成</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="new-task-form">
-          <label>リスト</label><br />
-          <select onChange={(e) => handleSelectList(e.target.value)} className="new-task-select-list">
+          <label>リスト</label>
+          <br />
+          <select
+            onChange={(e) => handleSelectList(e.target.value)}
+            className="new-task-select-list"
+          >
             {lists.map((list, key) => (
-              <option key={key} className="list-item" value={list.id}>{list.title}</option>
+              <option key={key} className="list-item" value={list.id}>
+                {list.title}
+              </option>
             ))}
-          </select><br />
-          <label>タイトル</label><br />
-          <input type="text" onChange={handleTitleChange} className="new-task-title" /><br />
-          <label>詳細</label><br />
-          <textarea type="text" onChange={handleDetailChange} className="new-task-detail" /><br />
-          <button type="button" className="new-task-button" onClick={onCreateTask}>作成</button>
+          </select>
+          <br />
+          <label>タイトル</label>
+          <br />
+          <input
+            type="text"
+            onChange={handleTitleChange}
+            className="new-task-title"
+          />
+          <br />
+          <label>詳細</label>
+          <br />
+          <textarea
+            type="text"
+            onChange={handleDetailChange}
+            className="new-task-detail"
+          />
+          <br />
+          <button
+            type="button"
+            className="new-task-button"
+            onClick={onCreateTask}
+          >
+            作成
+          </button>
         </form>
       </main>
     </div>
-  )
-}
+  );
+};
