@@ -1,18 +1,17 @@
 import React, { useState } from "react";
 import axios from "axios";
 import { useCookies } from "react-cookie";
-import { Redirect, useHistory, Link } from "react-router-dom";
+import { useNavigate, Link, Navigate } from "react-router-dom";
 import { Header } from "../components/Header";
 import "./signin.css";
 import { useDispatch, useSelector } from "react-redux";
 import { signIn } from "../authSlice";
 import { url } from "../const";
 
-
 export const SignIn = () => {
-  const auth = useSelector((state) => state.auth.isSignIn)
+  const auth = useSelector((state) => state.auth.isSignIn);
   const dispatch = useDispatch();
-  const history = useHistory();
+  const history = useNavigate();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [errorMessage, setErrorMessage] = useState();
@@ -20,7 +19,8 @@ export const SignIn = () => {
   const handleEmailChange = (e) => setEmail(e.target.value);
   const handlePasswordChange = (e) => setPassword(e.target.value);
   const onSignIn = () => {
-    axios.post(`${url}/signin`, {email: email, password: password})
+    axios
+      .post(`${url}/signin`, { email: email, password: password })
       .then((res) => {
         setCookie("token", res.data.token);
         dispatch(signIn());
@@ -28,26 +28,40 @@ export const SignIn = () => {
       })
       .catch((err) => {
         setErrorMessage(`サインインに失敗しました。${err}`);
-      })
-  }
+      });
+  };
 
-  if(auth) return <Redirect to="/" />
+  if (auth) return <Navigate to="/" />;
 
   return (
     <div>
-      <Header/>
+      <Header />
       <main className="signin">
         <h2>サインイン</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="signin-form">
-          <label className="email-label">メールアドレス</label><br />
-          <input type="email" className="email-input" onChange={handleEmailChange} /><br />
-          <label className="password-label">パスワード</label><br />
-          <input type="password" className="password-input" onChange={handlePasswordChange} /><br />
-          <button type="button" className="signin-button" onClick={onSignIn}>サインイン</button>
+          <label className="email-label">メールアドレス</label>
+          <br />
+          <input
+            type="email"
+            className="email-input"
+            onChange={handleEmailChange}
+          />
+          <br />
+          <label className="password-label">パスワード</label>
+          <br />
+          <input
+            type="password"
+            className="password-input"
+            onChange={handlePasswordChange}
+          />
+          <br />
+          <button type="button" className="signin-button" onClick={onSignIn}>
+            サインイン
+          </button>
         </form>
         <Link to="/signup">新規作成</Link>
       </main>
     </div>
-  )
-}
+  );
+};
