@@ -2,14 +2,14 @@ import axios from "axios";
 import React, { useState } from "react";
 import { useCookies } from "react-cookie";
 import { useSelector, useDispatch } from "react-redux";
-import { useHistory, Redirect } from "react-router-dom";
+import { useNavigate, Navigate } from "react-router-dom";
 import { signIn } from "../authSlice";
 import { Header } from "../components/Header";
 import { url } from "../const";
 import "./signUp.css";
 
 export const SignUp = () => {
-  const history = useHistory();
+  const history = useNavigate();
   const auth = useSelector((state) => state.auth.isSignIn);
   const dispatch = useDispatch();
   const [email, setEmail] = useState("");
@@ -24,10 +24,11 @@ export const SignUp = () => {
     const data = {
       email: email,
       name: name,
-      password: password
+      password: password,
     };
 
-    axios.post(`${url}/users`, data)
+    axios
+      .post(`${url}/users`, data)
       .then((res) => {
         const token = res.data.token;
         dispatch(signIn());
@@ -36,10 +37,10 @@ export const SignUp = () => {
       })
       .catch((err) => {
         setErrorMessge(`サインアップに失敗しました。 ${err}`);
-      })
+      });
 
-      if(auth) return <Redirect to="/" />
-  }
+    if (auth) return <Navigate to="/" />;
+  };
   return (
     <div>
       <Header />
@@ -47,15 +48,35 @@ export const SignUp = () => {
         <h2>新規作成</h2>
         <p className="error-message">{errorMessage}</p>
         <form className="signup-form">
-          <label>メールアドレス</label><br />
-          <input type="email" onChange={handleEmailChange} className="email-input" /><br />
-          <label>ユーザ名</label><br />
-          <input type="text" onChange={handleNameChange} className="name-input" /><br />
-          <label>パスワード</label><br />
-          <input type="password" onChange={handlePasswordChange} className="password-input" /><br />
-          <button type="button" onClick={onSignUp} className="signup-button">作成</button>
+          <label>メールアドレス</label>
+          <br />
+          <input
+            type="email"
+            onChange={handleEmailChange}
+            className="email-input"
+          />
+          <br />
+          <label>ユーザ名</label>
+          <br />
+          <input
+            type="text"
+            onChange={handleNameChange}
+            className="name-input"
+          />
+          <br />
+          <label>パスワード</label>
+          <br />
+          <input
+            type="password"
+            onChange={handlePasswordChange}
+            className="password-input"
+          />
+          <br />
+          <button type="button" onClick={onSignUp} className="signup-button">
+            作成
+          </button>
         </form>
       </main>
     </div>
-  )
-}
+  );
+};
