@@ -1,10 +1,11 @@
-import './App.css';
-import { Router } from './routes/Router';
+import "./App.css";
+import React from "react";
+import { Router } from "./routes/Router";
 
 function App() {
   return (
     <div className="App">
-        <Router />
+      <Router />
     </div>
   );
 }
